add SQLAlchemy based database backend
The database backend is a drop-in replacement for the MockDB used for
testing and development and implements the abstract IouDb interface.

To get the current design working - strict separation of internal model
classes and database schema mapper classes - the queries to the database
always join together all associated resources. The SQLAlchemy default
way is lazy-loading of database resources when accessing the properties
within a session.

Adding multiple schema objects with relationships to one another does
not behave intuitively: e.g. when adding a new group with existing users
in it SQLAlchemy always tries to do an INSERT instead of an "upsert". To
mitigate this one has to add existing user schema objects from the
database to the group object before adding it to the session for
insertion. It's also important, that this happens in the same session
such that the user schema object is associated with database entities
upon inerstion of the group schema object. Otherwise it would be treated
as a new object.

To convert database schema objects to internal lib objects the FastAPI
jsonable_encoder is required as the default dict encoder used by
pydantic does not correctly serialize SQLAlchemy schema classes to
dicts (The double * notation unpacks the dictionaries into key=value
pairs / kwargs expected by the pydantic constructor)