add pre-authenticated auth and prep authorization
Add abstract Authentication and Authorization classes. Implement the
pre-authenticated mechanism (currently used hard-coded).

Pre-authentication uses a proxy to do the authentication and take
control of the x-iou-pre-authenticated header, writing a valid username
into the x-iou-pre-authenticated header.

The currently authenticated user is not created automtatically, which is
still a TODO.

The abstract Authentication class can later be implemented using a token
parser for OIDC for example.

Authorization is planned via the abstract Authorization class.
Implementations of this class may parse authorization information from
an OIDC token or ask another service for authorization checks in the
forward-auth method.