lib/split: fall back to deposits as withdrawals
If no withdrawals are explicitly provided as split_parameters (which is
possible by providing an empty object in the API) assume deposit users
to be the withdrawals users under the equal split strategy. This is
mostly a convenience feature so you don't have to provide all user ids
in the parameters again if this special case is what you are aiming for.

Following up on this the API should now probably not treat the
split_parameters as mandatory. We'll have to evaluate if other split
strategies can also be equipped with sane defaults to omit treat the
params as optional in general.